+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Next
 
+* New user option `company-abort-on-unique-match`
+  ([#1046](https://github.com/company-mode/company-mode/issues/1046)).
 * `company-select-mouse` is a new frontend action
   ([#1045](https://github.com/company-mode/company-mode/pull/1045)).
 * `company-gtags` on remote hosts is improved